# backend/src/main/resources/contracts/openapi/zani.yaml
openapi: 3.0.3
info:
  title: ZANI API
  description: Google 로그인, 세션(강의방) 생성·입장·목록 조회 계약
  version: "1.0.0"
servers:
  - url: /api/v1

paths:
  /auth/login/google:
    post:
      operationId: loginWithGoogle
      summary: Google 계정으로 로그인
      description: |
        프론트에서 Google Identity Services로 발급받은 ID Token을 검증하고,
        최초 로그인이면 회원을 생성한 뒤 자체 access/refresh 토큰을 발급한다.
        refresh 토큰은 응답 바디가 아닌 HttpOnly 쿠키(auth/refresh 경로)로 내려간다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginRequest"
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleLoginApiResponse"
        "400":
          description: 요청 검증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Google ID Token이 유효하지 않음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: AUTH_008
                message: Google ID token is invalid

  /sessions:
    post:
      operationId: createSession
      summary: 세션(강의방) 생성
      description: |
        진행 중인 방이 없는 강사가 제목과 함께 세션을 생성한다.
        PREPARING 상태로 만들어지고 고유 초대 코드가 함께 발급되지만,
        강사가 `POST /sessions/{sessionId}/start` 를 호출하기 전까지 그 코드로는 입장할 수 없다.
        아직 시작하지 않았으므로 `expiresAt` 은 null 이다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: 세션 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: 이미 진행 중인 세션이 있음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: SESSION_APP_001
                message: An active session already exists

    get:
      operationId: listSessions
      summary: 내가 참여 중인 세션 목록 조회
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 세션 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /sessions/{sessionId}/start:
    post:
      operationId: startSession
      summary: 준비된 세션을 실제로 시작
      description: |
        PREPARING 세션을 LIVE 로 전환한다. 이 시점부터 초대 코드가 유효해지고
        3시간 자동 종료 시계가 돌기 시작한다(`expiresAt`).

        **강사가 미디어 서버에 연결된 뒤 호출한다.** 연결 전에 시작하면 연결이 실패했을 때
        초대 코드만 열려 학생이 강사 없는 방에 들어온다.

        멱등하다. 이미 LIVE 인 세션에 다시 호출하면 시작 시각을 밀지 않고 `started: false` 로 응답한다.
        동시 요청은 세션 행 잠금으로 직렬화되어 한 건만 `started: true` 가 된다.
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 세션 시작 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartSessionApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: 이 세션을 연 강사가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: SESSION_APP_006
                message: Only the instructor who opened this session can control it
        "404":
          description: 그런 세션이 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: SESSION_APP_005
                message: No session found
        "409":
          description: 이미 종료 절차에 들어간 세션
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: SESSION_002
                message: Session is no longer live

  /sessions/joinable:
    post:
      operationId: checkJoinable
      summary: 초대 코드로 들어갈 수 있는지 확인
      description: |
        초대 코드가 유효하고 지금 들어갈 수 있는지만 확인한다. **참가자를 만들지 않는다.**

        입장 전 점검 화면으로 넘어가기 전에 쓴다. 이 확인을 `/sessions/join` 으로 대신하면
        참가자 행이 먼저 생겨 정원을 선점한다 — 학생이 장치 점검에서 이탈해도 자리가 반환되지 않아,
        그런 학생이 29명이면 실제 입장자 없이 방이 찬다.

        거절 사유와 오류 코드는 `/sessions/join` 과 같다. `remainingSeats` 는 확인한 순간의
        값이라 실제 입장까지 줄어들 수 있다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinSessionRequest"
      responses:
        "200":
          description: 입장 가능
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinableSessionApiResponse"
        "400":
          description: 정규화 후에도 8자가 아닌 초대 코드
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: 유효하지 않은 초대 코드
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: SESSION_APP_005
                message: No session found for the given invite code
        "409":
          description: 아직 시작하지 않았거나(SESSION_APP_007) 이미 끝났거나(SESSION_APP_008) 정원이 찬(SESSION_APP_009) 세션
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

  /sessions/join:
    post:
      operationId: joinSession
      summary: 초대 코드로 세션 입장
      description: |
        유효한 초대 코드로 입장하면 STUDENT 멤버십을 즉시 얻는다.
        이미 참여 중인 세션이면 멱등하게 기존 멤버십을 반환한다.

        코드는 서버가 정규화한다. 하이픈·공백을 지우고 대문자로 올린 뒤 8자인지 본다.
        입장은 LIVE 세션에만 열린다 — 아직 시작하지 않았거나 이미 끝난 세션은 409 다.
        정원은 강사를 포함해 30명이며, 이미 들어와 있는 참가자의 재입장은 정원을 다시 쓰지 않는다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinSessionRequest"
      responses:
        "200":
          description: 세션 입장 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinSessionApiResponse"
        "400":
          description: 정규화 후에도 8자가 아닌 초대 코드
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: 유효하지 않은 초대 코드
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: SESSION_APP_005
                message: No session found for the given invite code
        "409":
          description: 아직 시작하지 않았거나(SESSION_APP_007) 이미 끝났거나(SESSION_APP_008) 정원이 찬(SESSION_APP_009) 세션
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                isSuccess: false
                code: SESSION_APP_009
                message: Session has reached its participant limit

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    Unauthorized:
      description: 인증되지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
          example:
            isSuccess: false
            code: COMM_401
            message: Authentication is required

  schemas:
    GoogleLoginRequest:
      type: object
      properties:
        idToken:
          type: string
      required:
        - idToken

    GoogleLoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        email:
          type: string
        displayName:
          type: string
        profileImageUrl:
          type: string
          nullable: true
        newMember:
          type: boolean

    GoogleLoginApiResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        data:
          $ref: "#/components/schemas/GoogleLoginResponse"
      required:
        - isSuccess

    # 확정 생명주기. PREPARING → LIVE → ENDING → NOTE_PENDING → ENDED 로만 진행하고 역행하지 않는다.
    # 학생 입장과 접속 집계는 LIVE 에서만 열린다.
    SessionStatus:
      type: string
      enum:
        - PREPARING
        - LIVE
        - ENDING
        - NOTE_PENDING
        - ENDED

    SessionRole:
      type: string
      enum:
        - INSTRUCTOR
        - STUDENT

    # 공통 응답 봉투. data 형태가 없는 에러 응답 등 범용으로 쓴다.
    ApiResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        data:
          nullable: true
      required:
        - isSuccess

    CreateSessionRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
      required:
        - title

    CreateSessionResponse:
      type: object
      properties:
        sessionId:
          # TSID 는 JavaScript 안전 정수 상한(2^53-1)을 넘어 브라우저에서 값이 반올림된다.
          # 그래서 식별자는 문자열로 내보낸다.
          type: string
        inviteCode:
          type: string
          minLength: 8
          maxLength: 8
        status:
          $ref: "#/components/schemas/SessionStatus"
        role:
          $ref: "#/components/schemas/SessionRole"
        # 아직 시작하지 않은 PREPARING 세션은 자동 종료 시계가 돌지 않아 null 이다.
        expiresAt:
          type: string
          format: date-time
          nullable: true

    CreateSessionApiResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        data:
          $ref: "#/components/schemas/CreateSessionResponse"
      required:
        - isSuccess

    StartSessionResponse:
      type: object
      properties:
        sessionId:
          # TSID 는 JavaScript 안전 정수 상한(2^53-1)을 넘어 브라우저에서 값이 반올림된다.
          # 그래서 식별자는 문자열로 내보낸다.
          type: string
        inviteCode:
          type: string
          minLength: 8
          maxLength: 8
        status:
          $ref: "#/components/schemas/SessionStatus"
        expiresAt:
          type: string
          format: date-time
          nullable: true
        # 이번 요청으로 시작됐으면 true, 이미 진행 중이었으면 false.
        started:
          type: boolean

    StartSessionApiResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        data:
          $ref: "#/components/schemas/StartSessionResponse"
      required:
        - isSuccess

    JoinSessionRequest:
      type: object
      properties:
        # 서버가 하이픈·공백을 지우고 대문자로 올린 뒤 8자를 검증한다.
        # 원문 길이를 8자로 못 박으면 "ABCD-EFGH" 같은 표기가 형식 오류로 막힌다.
        inviteCode:
          type: string
          minLength: 8
          maxLength: 20
      required:
        - inviteCode

    JoinableSessionResponse:
      type: object
      properties:
        sessionId:
          # TSID 는 JavaScript 안전 정수 상한(2^53-1)을 넘어 브라우저에서 값이 반올림된다.
          type: string
        inviteCode:
          type: string
        status:
          $ref: "#/components/schemas/SessionStatus"
        # 확인한 순간의 남은 자리. 실제 입장까지 줄어들 수 있어 표시용이다.
        remainingSeats:
          type: integer

    JoinableSessionApiResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        data:
          $ref: "#/components/schemas/JoinableSessionResponse"
      required:
        - isSuccess

    JoinSessionResponse:
      type: object
      properties:
        sessionId:
          # TSID 는 JavaScript 안전 정수 상한(2^53-1)을 넘어 브라우저에서 값이 반올림된다.
          # 그래서 식별자는 문자열로 내보낸다.
          type: string
        inviteCode:
          type: string
        status:
          $ref: "#/components/schemas/SessionStatus"
        role:
          $ref: "#/components/schemas/SessionRole"

    JoinSessionApiResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        data:
          $ref: "#/components/schemas/JoinSessionResponse"
      required:
        - isSuccess

    SessionSummary:
      type: object
      properties:
        sessionId:
          # TSID 는 JavaScript 안전 정수 상한(2^53-1)을 넘어 브라우저에서 값이 반올림된다.
          # 그래서 식별자는 문자열로 내보낸다.
          type: string
        inviteCode:
          type: string
        status:
          $ref: "#/components/schemas/SessionStatus"
        role:
          $ref: "#/components/schemas/SessionRole"

    SessionListApiResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
      required:
        - isSuccess